+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -121,6 +121,11 @@ New checks
   Detects issues with regards to AUTOSAR types and their associated macro values
   by essentially treating these types as enumerations.
 
+- New :doc:`embedded-no-early-return
+  <clang-tidy/checks/embedded/no-early-return>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`misc-no-comma-operator
   <clang-tidy/checks/misc/no-comma-operator>` check.